requirements.txt
Box2D==2.3.10
gym==0.21.0
keras==2.10.0
matplotlib==3.6.0
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.4.0
pyglet==1.5.27
pypng==0.20220715.0
scikit_image==0.19.3
scipy==1.9.1
skimage==0.0
